+++ Doctor_Login_Registry/src/main/java/com/hospital/DoctorLoginRegistryApplication.java
@@ -1,4 +1,4 @@
-package com.patient;
+package com.hospital;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
