+++ Doctor_Login_Registry/src/main/java/com/hospital/entity/Doctor.java
@@ -1,9 +1,9 @@
-package com.patient.entity;
+package com.hospital.entity;
 
-import org.springframework.data.annotation.Id;
 
-import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
+import jakarta.persistence.Column;
+import jakarta.persistence.Id;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import lombok.AllArgsConstructor;
