+++ Doctor_Login_Registry/src/main/java/com/hospital/repository/Doctor_InterfaceRepository.java
@@ -1,11 +1,12 @@
-package com.patient.repository;
+package com.hospital.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.patient.entity.Doctor;
+import com.hospital.entity.Doctor;
 
 @Repository
 public interface Doctor_InterfaceRepository extends JpaRepository<Doctor, Long>{
-
+	
+	Doctor findByUserName(String userName);
 }
