+++ Doctor_Login_Registry/src/main/java/com/hospital/service/Doctor_InterfaceService.java
@@ -1,0 +1,9 @@
+package com.hospital.service;
+
+import com.hospital.entity.Doctor;
+
+public interface Doctor_InterfaceService {
+
+	Boolean addDoctor(Doctor doctor);
+
+}
