--- Doctor_Login_Registry/src/main/java/com/patient/service/Doctor_InterfaceService.java
@@ -1,11 +1,0 @@
-package com.patient.service;
-
-import com.patient.entity.Doctor;
-
-public interface Doctor_InterfaceService {
-
-	void loginDoctor(Long id);
-
-	void addDoctor(Doctor doctor);
-
-}
