+++ Doctor_Login_Registry/src/main/resources/application.properties
@@ -1,1 +1,10 @@
+server.port= 8081
 
+spring.datasource.url=jdbc:mysql://localhost:3306/HospitalDatabase
+spring.datasource.username= root
+spring.datasource.password= rs123
+spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
+
+spring.jpa.properties.hibernate.dialect= org.hibernate.dialect.MySQLDialect
+spring.jpa.hibernate.ddl-auto= update
+spring.jpa.show-sql= true
